requirements.txt
aiogram
requests
python-dotenv
fastapi
uvicorn
pytest
httpx